Consertando os problemas encontrados na auditoria

+++ excluiAluno.php
@@ -2,7 +2,7 @@
 session_start();
 include_once './valida_login.php';
 /* Attempt MySQL server connection. Assuming you are running MySQL
-  server with default setting (user 'root' with no password) */
+server with default setting (user 'root' with no password) */
 include_once './mysql.php';
 
 include_once './topo.php';
@@ -11,19 +11,38 @@ include_once './topo.php';
 // Attempt insert query execution
 try {
     // Create prepared statement
-    $id = addslashes($_POST['id_aluno']);
+	$id = addslashes($_POST['id_aluno']);
 
-    $sql = "DELETE FROM aluno WHERE id_aluno = $id";
+	$sql = "SELECT * FROM aluno WHERE id_aluno = $id";
 
-    $stmt = $pdo->prepare($sql);
+	$stmt = $pdo->prepare($sql);
 
-    $stmt->execute();
+	$stmt->execute();
+
+	$arrValues = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
 
-    echo "<script type=text/javascript>alert('Operação realizada com sucesso!');window.location='removeAluno.php'</script>";
+	if (empty($arrValues)) {
+		?>    
+		echo "<script type=text/javascript>alert('Aluno registro encontrado!');window.location='removeAluno.php'</script>";
+
+		<?php
+	} else {
+
+
+		$sql = "DELETE FROM aluno WHERE id_aluno = $id";
+
+		$stmt = $pdo->prepare($sql);
+
+		$stmt->execute();
+
+
+		echo "<script type=text/javascript>alert('Operação realizada com sucesso!');window.location='removeAluno.php'</script>";
+
+	}
 } catch (PDOException $e) {
     //die("ERROR: Could not able to execute $sql. " . $e->getMessage());
-    echo "<script type=text/javascript>alert('Operação NÃO realizada!');window.location='removeAluno.php'</script>";
+	echo "<script type=text/javascript>alert('Operação NÃO realizada!');window.location='removeAluno.php'</script>";
 }
 
 // Close connection
